11.24

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Post;
+use  App\Comment;
 
 class PostController extends Controller
 {
@@ -11,12 +12,14 @@ class PostController extends Controller
     public function index(){
 
         $posts = Post::orderBy('created_at','desc')
-            ->paginate(6);
+           ->withCount('comments') ->paginate(6);
         return view("post/index" ,compact('posts'));
     }
 
     //文章详情
     public function show(Post $post){
+
+        $post->load('comments');
         return view("post/show",compact('post'));
     }
 
@@ -33,8 +36,15 @@ class PostController extends Controller
             'content' => 'required|string|min:10',
         ]);
 
-        $post = Post::create(request(['title','content']));
 
+
+        //逻辑
+        $user_id = \Auth::id();
+        $params = array_merge(request(['title','content']),compact('user_id'));
+
+        $post = Post::create($params );
+
+        //渲染
         return redirect("/posts");
     }
 
@@ -52,6 +62,7 @@ class PostController extends Controller
             'content' => 'required|string|min:10',
         ]);
 
+        $this->authorize('update',$post);
         //逻辑
         $post->title = request('title');
         $post->content = request('content');
@@ -63,7 +74,7 @@ class PostController extends Controller
 
     //删除逻辑
     public function delete(Post $post){
-        //TODO:用户权限认证
+        $this->authorize('delete',$post);
         $post->delete();
         return redirect("/posts");
     }
@@ -73,4 +84,20 @@ class PostController extends Controller
         $path =  $request->file('wangEditorH5File')->storePublicly(md5(time()));
         return asset('storage/'.$path);
     }
+
+    //提交评论
+    public function comment(Post $post){
+        $this->validate(request(),[
+            'content' => 'required|min:3',
+        ]);
+
+        //逻辑
+        $comment = new Comment();
+        $comment->user_id = \Auth::id();
+        $comment->content = request('content');
+        $post->comments()->save($comment);
+
+        return back();
+    }
+
 }
